// src/app/modules/reactive-forms/src/lib/components/rf-input-text/enums/rf-autocomplete-types.enum.ts
export enum AutocompleteTypes {
  // Personal information
  NAME = 'given-name',
  LAST_NAME = 'family-name',
  CITY = 'address-level2',
  ADDRESS = 'street-address',
  POSTAL_CODE = 'postal-code',
  // Contact information
  COMPLETE_PHONE = 'tel',
  PHONE = 'tel-national',
  EMAIL = 'email',
  // Generic
  OFF = 'off',
  ON = 'on',
}