// src/app/modules/reactive-forms/src/lib/components/rf-input-text/models/rf-input-types.model.ts
export enum RfInputTypes {
  TEXT = 'text',
  PASSWORD = 'password',
  EMAIL = 'email',
  NUMBER = 'number',
  TEL = 'tel',
  URL = 'url',
  SEARCH = 'search',
  DATE = 'date',
  TIME = 'time',
  DATETIME_LOCAL = 'datetime-local',
  MONTH = 'month',
  WEEK = 'week',
  // CHECKBOX = 'checkbox',
  // RADIO = 'radio',
  FILE = 'file',
  RANGE = 'range',
  COLOR = 'color',
  HIDDEN = 'hidden',
  SUBMIT = 'submit',
  RESET = 'reset',
  BUTTON = 'button',
  IMAGE = 'image',
}