// src/app/modules/reactive-forms/src/lib/form-builder/enums/rf-form-builder.types.enum.ts
export enum RfFormBuilderFieldType {
  HTML_INJECTION = 'rf-html-injection',
  INPUT = 'rf-input',
  RADIO = 'rf-radio',
  CHECKBOX = 'rf-checkbox',
  SWITCH = 'rf-switch',
  LIST = 'rf-list',
  SELECT = 'rf-select',
  IP = 'rf-ip-input',
  PREFIX_PHONE = 'rf-prefix-phone',
  DATEPICKER = 'rf-datepicker',
  INPUT_DATEPICKER = 'rf-input-datepicker',
  SELECT_DATE_PICKER = 'rf-select-date-picker',
}